requirements.txt
Flask
Pillow
pathlib
pdf2image
python-magic
sanic
sanic_openapi
tesserocr
